refactor : dedup test fixtures + share TestConfig load across read-only files
Merge the three read-only TestConfig page-structure files (controllers/auth,
controllers/select, zones/structure) into one test_page_structure_e2e.py
sharing a single module fixture. Move the Possession OR-shape regression
into test_vampire_anchor_zones_e2e.py so it stops paying for an in-class
Vampire1966CSV swap-in plus a swap-back to TestConfig. Add
ui_controller_ids_map() in helpers.py and route the five hand-rolled
controller-id scrapers through it. Add ensure_scenario_loaded() that skips
the admin reset when the same scenario is already current, used by the
three TestConfig worker files.
