+++ tests/test_attack_location_baseline_e2e.py
@@ -51,6 +51,7 @@ from conftest import (
 from helpers import (
     DB_AVAILABLE, end_turn, load_minimal_data, load_scenario_via_admin, safe_goto,
     register_php_error_listener, assert_no_collected_php_errors,
+    ui_controller_id,
 )
 
 
@@ -90,14 +91,7 @@ def _location_id_via_management(page, location_name):
 
 
 def _controller_id_via_management(page, controller_lastname):
-    safe_goto(page, f"{PHP_BASE_URL}/base/accueil.php")
-    page.wait_for_load_state("load")
-    for opt in page.locator("select#controllerSelect option").all():
-        v = opt.get_attribute("value") or ""
-        t = (opt.inner_text() or "").strip()
-        if v and controller_lastname in t:
-            return int(v)
-    raise AssertionError(f"controller_id for '{controller_lastname}' not found")
+    return ui_controller_id(page, controller_lastname, PHP_BASE_URL)
 
 
 def _seed_ckl_admin(page, controller_lastname, location_name):
