+++ tests/test_transformation_unlock_rules_e2e.py
@@ -582,56 +582,3 @@ class TestAdminBypass:
             f"gold before={gold_before} after={_read_amount(alpha, gold)}"
         )
         _set_amount(alpha, gold, 100)
-
-
-# ---------------------------------------------------------------------------
-# R10 — Vampire1966 Possession OR regression
-# ---------------------------------------------------------------------------
-
-class TestVampire1966PossessionOR:
-    """Regression check for R10: the Possession row's pre-fix OR was a
-    single object that PHP iterated as scalar keys → $OR stayed false →
-    power fully hidden. Post-fix it is an array of two single-key
-    branches matching D12 (array-of-objects) + D20 (one rule per branch).
-
-    Belt-and-buckle DB check: read the Possession row, decode its
-    on_transformation.OR, assert it is a list of 2 single-key dicts."""
-
-    @pytest.fixture(scope="class", autouse=True)
-    def vampire_scenario(self, browser):
-        load_scenario_via_admin(browser, PHP_BASE_URL, "Vampire1966CSV")
-        yield
-        load_scenario_via_admin(browser, PHP_BASE_URL, "TestConfig")
-
-    def test_possession_or_shape_is_array_of_single_key_objects(self):
-        import json as _json
-        conn = _db(); cur = conn.cursor()
-        cur.execute(
-            f"SELECT other FROM `{GAME_PREFIX}powers` WHERE name='Possession' LIMIT 1"
-        )
-        row = cur.fetchone()
-        cur.close(); conn.close()
-        assert row is not None, "Possession row not loaded from Vampire1966CSV"
-
-        other = _json.loads(row['other'])
-        ot = other.get('on_transformation') or {}
-        or_block = ot.get('OR')
-
-        assert isinstance(or_block, list), (
-            f"D12: OR must be array-of-objects, got {type(or_block).__name__}: {or_block}"
-        )
-        assert len(or_block) == 2, (
-            f"Possession fix locks OR as TWO single-key branches; got {or_block}"
-        )
-        for branch in or_block:
-            assert isinstance(branch, dict), (
-                f"D20: each OR branch must be an object; got {type(branch).__name__}"
-            )
-            assert len(branch) == 1, (
-                f"R10 fix authoring convention: one key per branch; got {branch}"
-            )
-        keys = {next(iter(b.keys())) for b in or_block}
-        assert keys == {"age", "worker_is_alive"}, (
-            f"Possession OR must carry exactly the age and worker_is_alive "
-            f"branches; got keys {keys}"
-        )
