+++ tests/test_workers_mass_actions_e2e.py
@@ -22,7 +22,7 @@ from playwright.sync_api import Page
 
 from conftest import PHP_BASE_URL, ensure_gm_login
 from helpers import (
-    DB_AVAILABLE, load_minimal_data, load_scenario_via_admin, login_as, safe_goto,
+    DB_AVAILABLE, load_minimal_data, ensure_scenario_loaded, login_as, safe_goto,
     register_php_error_listener, assert_no_collected_php_errors,
     ui_worker_id, ui_all_workers,
     ui_mass_hide_click, ui_mass_investigate_click, ui_mass_passive_click,
@@ -39,10 +39,10 @@ def base_url():
 
 @pytest.fixture(scope="module", autouse=True)
 def setup_testconfig(browser):
-    """Load TestConfig once per module."""
+    """Load TestConfig once per module (skipped if already loaded)."""
     if DB_AVAILABLE:
         load_minimal_data()
-    load_scenario_via_admin(browser, PHP_BASE_URL, "TestConfig")
+    ensure_scenario_loaded(browser, PHP_BASE_URL, "TestConfig")
     yield
